refactor `Footer` index block
